refactor: Split Models.swift into separate files for better organization
- Created individual files for each model: Speaker, Session, Schedule, SponsorItem, SponsorGroup, Partner, SponsorsData, and Staff
- Moved shared URL decoding extension to URLDecodingExtension.swift
- Applied swift-format to all files
- All tests passing
