+++ Sources/SessionData/Schedule.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct Schedule: Codable {
+  public let day1: [Session]
+  public let day2: [Session]
+}
