+++ Sources/SessionData/SponsorGroup.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct SponsorGroup: Codable {
+  public let items: [SponsorItem]
+  public let title: String
+}
