+++ Sources/SessionData/SponsorItem.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+public struct SponsorItem: Codable {
+  public let name: String
+  public let picture: URL?
+  public let link: URL?
+
+  public init(from decoder: Decoder) throws {
+    let container = try decoder.container(keyedBy: CodingKeys.self)
+
+    name = try container.decode(String.self, forKey: .name)
+
+    // Decode URL fields using shared extension
+    picture = try container.decodeURL(forKey: .picture)
+    link = try container.decodeURL(forKey: .link)
+  }
+}
