+++ Sources/SessionData/SponsorsData.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct SponsorsData: Codable {
+  public let sponsors: [SponsorGroup]
+  public let partner: [Partner]
+}
